Bugfix: TypeError: client.sourceContents is not a function
